requirements.txt
xai-sdk
openai

# Core dependencies
pydantic>=2.0.0
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0

# Export functionality
reportlab>=4.0.0
markdown>=3.4.0
jinja2>=3.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
